Moving CSS require logic into seperate package
