+++ workspaces/templates/packages/server-side-rendering/package.json
@@ -34,8 +34,6 @@
     "@goldstack/template-ssr-cli": "0.1.2",
     "@goldstack/utils-aws-http-api-local": "0.1.34",
     "@jest-mock/express": "^1.4.5",
-    "@swc-node/register": "^1.5.1",
-    "@swc/jest": "^0.2.22",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
     "@types/aws-lambda": "^8.10.88",
@@ -49,10 +47,8 @@
     "find-free-port": "^2.0.0",
     "jest": "^28.1.0",
     "jest-environment-jsdom": "^28.1.3",
+    "node-css-require": "0.1.0",
     "node-fetch": "2.6.7",
-    "pirates": "^4.0.5",
-    "postcss": "^8.4.14",
-    "postcss-modules-sync": "^1.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "rimraf": "^3.0.2",
