+++ workspaces/templates/packages/server-side-rendering/tsconfig.json
@@ -17,7 +17,9 @@
     "moduleResolution": "node",
     "isolatedModules": true,
     "plugins": [
-      { "name": "typescript-plugin-css-modules" }
+      {
+        "name": "typescript-plugin-css-modules"
+      }
     ]
   },
   "exclude": [
@@ -35,6 +37,9 @@
     },
     {
       "path": "../../../templates-lib/packages/utils-aws-http-api-local"
+    },
+    {
+      "path": "../../../utils/packages/node-css-require"
     }
-  ],
+  ]
 }
