+++ workspaces/utils/packages/node-css-require/package.json
@@ -1,0 +1,53 @@
+{
+  "name": "node-css-require",
+  "version": "0.1.0",
+  "description": "Load CSS modules when running Node.js CLI and server applications",
+  "keywords": [
+    "node",
+    "javascript",
+    "css"
+  ],
+  "homepage": "",
+  "bugs": {
+    "url": "https://github.com/goldstack/goldstack/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/goldstack/goldstack.git"
+  },
+  "license": "MIT",
+  "author": "Max Rohde",
+  "sideEffects": false,
+  "main": "src/nodeCssRequire.ts",
+  "scripts": {
+    "build": "yarn clean && yarn compile",
+    "build:watch": "yarn clean && yarn compile-watch",
+    "clean": "rimraf ./dist",
+    "compile": "tsc -p tsconfig.json",
+    "compile-watch": "tsc -p tsconfig.json --watch",
+    "compile-watch:light": "nodemon --watch ./src/ -e '*' --exec 'yarn compile'",
+    "coverage": "jest --collect-coverage --passWithNoTests --config=./jest.config.js --runInBand",
+    "prepublishOnly": "yarn run build",
+    "publish": "utils-git changed --exec \"yarn npm publish $@\"",
+    "test-ci": "jest --passWithNoTests --config=./jest.config.js --runInBand",
+    "version:apply": "utils-git changed --exec \"yarn version $@ && yarn version apply\""
+  },
+  "dependencies": {
+    "@swc-node/register": "^1.5.1",
+    "pirates": "^4.0.5",
+    "postcss": "^8.4.14",
+    "postcss-modules-sync": "^1.0.0"
+  },
+  "devDependencies": {
+    "@goldstack/utils-git": "0.1.36",
+    "@types/jest": "^27.5.1",
+    "@types/node": "^17.0.33",
+    "jest": "^28.1.0",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^28.0.2",
+    "typescript": "^4.7.4"
+  },
+  "publishConfig": {
+    "main": "dist/src/nodeCssRequire.js"
+  }
+}
